refactor(sandbox): split attribution.rs test module into its own file
My prior fixes pushed attribution.rs from 951 to 1029 lines, crossing the
hard 1000-line threshold. Moved #[cfg(test)] mod tests (FakeLookup harness,
17 unit tests, the gated real-Docker integration test) into a sibling
attribution_tests.rs via #[path], matching the file's existing docker_gate
#[path] pattern. Pure move, no behavior change: production attribution.rs
is now 367 lines.
